@@ -29,11 +29,11 @@
   },
   "minimum-stability": "stable",
   "require": {
-    "php": "^7.1",
+    "php": "^7.4 || ^8.0",
     "ext-mbstring": "*",
     "ext-openssl": "*",
-    "phpseclib/phpseclib": "^2.0",
-    "lcobucci/jwt": "3.3.3",
+    "phpseclib/phpseclib": "^3.0",
+    "lcobucci/jwt": "^4.0",
     "guzzlehttp/guzzle": "^6.0|^7.0"
   },
   "require-dev": {
