@@ -2,54 +2,43 @@
 
 namespace Azimo\Apple\Auth\Jwt;
 
-use Azimo\Apple\Api\AppleApiClient;
+use Azimo\Apple\Api\AppleApiClientInterface;
 use Azimo\Apple\Api\Exception as ApiException;
 use Azimo\Apple\Api\Response\JsonWebKeySet;
 use Azimo\Apple\Auth\Exception;
-use BadMethodCallException;
 use Lcobucci\JWT;
 use OutOfBoundsException;
-use phpseclib\Crypt\RSA;
-use phpseclib\Math\BigInteger;
+use phpseclib3\Crypt\RSA;
+use phpseclib3\Math\BigInteger;
 
 class JwtVerifier
 {
-    /**
-     * @var AppleApiClient
-     */
-    private $client;
+    private AppleApiClientInterface $client;
 
-    /**
-     * @var RSA
-     */
-    private $rsa;
+    private JWT\Signer $signer;
 
-    /**
-     * @var JWT\Signer
-     */
-    private $signer;
+    private JWT\Validator $validator;
 
-    public function __construct(AppleApiClient $client, RSA $rsa, JWT\Signer $signer)
+    public function __construct(AppleApiClientInterface $client, JWT\Validator $validator, JWT\Signer $signer)
     {
         $this->client = $client;
-        $this->rsa = $rsa;
         $this->signer = $signer;
+        $this->validator = $validator;
     }
 
     /**
      * @throws Exception\InvalidCryptographicAlgorithmException
      * @throws Exception\KeysFetchingFailedException
-     * @throws Exception\NotSignedTokenException
      */
     public function verify(JWT\Token $jwt): bool
     {
-        $this->loadRsaKey($this->getAuthKey($jwt));
-
-        try {
-            return $jwt->verify($this->signer, $this->rsa->getPublicKey());
-        } catch (BadMethodCallException $exception) {
-            throw  new Exception\NotSignedTokenException($exception->getMessage(), $exception->getCode(), $exception);
-        }
+        return $this->validator->validate(
+            $jwt,
+            new JWT\Validation\Constraint\SignedWith(
+                $this->signer,
+                JWT\Signer\Key\InMemory::plainText($this->createPublicKey($this->getAuthKey($jwt)))
+            )
+        );
     }
 
     /**
@@ -69,7 +58,7 @@ class JwtVerifier
         }
 
         try {
-            $cryptographicAlgorithm = $jwt->getHeader('kid');
+            $cryptographicAlgorithm = $jwt->headers()->get('kid');
             $authKey = $authKeys->getByCryptographicAlgorithm($cryptographicAlgorithm);
         } catch (OutOfBoundsException | ApiException\UnsupportedCryptographicAlgorithmException $exception) {
             throw new Exception\InvalidCryptographicAlgorithmException(
@@ -88,24 +77,11 @@ class JwtVerifier
         return $authKey;
     }
 
-    private function loadRsaKey(JsonWebKeySet $authKey): void
+    private function createPublicKey(JsonWebKeySet $authKey): string
     {
-        /**
-         * Phpspeclib is parsing phpinfo(); output to determine OpenSSL Library and Header versions,
-         * basing on that set if MATH_BIGINTEGER_OPENSSL_ENABLED or MATH_BIGINTEGER_OPENSSL_DISABLED const.
-         * It crashes tests so it is possible that it might crash production, that is why constants are overwritten.
-         *
-         * @see vendor/phpseclib/phpseclib/phpseclib/Math/BigInteger.php:273
-         */
-        if (!defined('MATH_BIGINTEGER_OPENSSL_ENABLED')) {
-            define('MATH_BIGINTEGER_OPENSSL_ENABLED', true);
-        }
-
-        $this->rsa->loadKey(
-            [
-                'exponent' => new BigInteger(base64_decode(strtr($authKey->getExponent(), '-_', '+/')), 256),
-                'modulus'  => new BigInteger(base64_decode(strtr($authKey->getModulus(), '-_', '+/')), 256),
-            ]
+        return RSA\Formats\Keys\PKCS8::savePublicKey(
+            new BigInteger(base64_decode(strtr($authKey->getModulus(), '-_', '+/')), 256),
+            new BigInteger(base64_decode(strtr($authKey->getExponent(), '-_', '+/')), 256)
         );
     }
 }
