@@ -8,12 +8,9 @@ use Azimo\Apple\Api\Factory\ResponseFactory;
 use Azimo\Apple\Api\Response\JsonWebKeySetCollection;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 
-class ResponseFactoryTest extends MockeryTestCase
+final class ResponseFactoryTest extends MockeryTestCase
 {
-    /**
-     * @var ResponseFactory
-     */
-    private $responseFactory;
+    private ResponseFactory $responseFactory;
 
     protected function setUp(): void
     {
