@@ -4,6 +4,6 @@ namespace Azimo\Apple\Auth\Exception;
 
 use InvalidArgumentException;
 
-class InvalidJwtException extends InvalidArgumentException implements AppleExceptionInterface
+final class InvalidJwtException extends InvalidArgumentException implements AppleExceptionInterface
 {
 }
