@@ -4,6 +4,6 @@ namespace Azimo\Apple\Auth\Exception;
 
 use InvalidArgumentException;
 
-class ValidationFailedException extends InvalidArgumentException implements AppleExceptionInterface
+final class ValidationFailedException extends InvalidArgumentException implements AppleExceptionInterface
 {
 }
